+++ lib/allegro/ffi_allegro_flac.rb
@@ -7,7 +7,7 @@ module Allegro extend FFI::Library
     attach_function :al_init_flac_addon, [], :bool
     attach_function :al_load_sample_flac, [:string], :pointer
     attach_function :al_load_audio_stream_flac, [:string,
-                    :uint, :uint], :pointer
+                    :size_t, :uint], :pointer
     attach_function :al_get_allegro_flac_version, [], :uint32
 end
